pyproject.toml
[project]
name = "sparrow-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "moderngl>=5.12.0",
    "numpy>=2.4.0",
    "pygame-ce>=2.5.6",
    "pytest>=9.0.2",
    "scipy>=1.16.3",
]

[tool.uv.workspace]
members = [
    "temp",
]
